[FIX] (chat) 중복코드 리팩토링 & 서비스 계층 구현과 인터페이스로 분리
