+++ src/main/java/com/example/lastproject/domain/chat/controller/WebSocketEventListener.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.socket.messaging.SessionConnectEvent;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
 import java.util.Objects;
+import java.util.Optional;
 
 /**
  * 소켓의 연결과 해제를 감지하기위한 이벤트리스너를 사용하는 클래스
@@ -35,24 +36,13 @@ public class WebSocketEventListener {
         // StompHeaderAccessor를 사용하여 세션 정보 접근
         StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
 
-        //JwtInterceptor에서 검증한 뒤 헤더에 저장해놓은 유저객체에서 email 추출
-        AuthUser authUser = (AuthUser) Objects.requireNonNull(headerAccessor.getSessionAttributes()).get("authUser");
-        String email = authUser.getEmail();
+        String email = getEmailFromSession(headerAccessor);
+        Long chatRoomId = getChatRoomIdFromHeader(headerAccessor);
 
-        //javaScript에서 헤더로 받아올수 있게 설정한 chatRoomId 추출
-        long chatRoomId = Long.parseLong(Objects.requireNonNull(headerAccessor.getFirstNativeHeader("chatRoomId")));
-
-        // 세션 속성에 chatRoomId 저장(아래 SessionDisconnectEvent에서 사용하기 위함. 저장해놓지 않으면 SessionDisconnectEvent에서 사용할 수 없음)
-        headerAccessor.getSessionAttributes().put("chatRoomId", chatRoomId);
-
-        //메세지 전송
-        if (email != null) {
-
-            ChatMessageRequest chatMessage = new ChatMessageRequest();
-            chatMessage.changeType(ChatMessageRequest.MessageType.JOIN); // JOIN 타입 설정
-            chatMessage.changeSender(email);
-
-            messagingTemplate.convertAndSend("/topic/" + chatRoomId, chatMessage);
+        if (email != null && chatRoomId != null) {
+            // 세션 속성에 chatRoomId 저장(아래 SessionDisconnectEvent에서 사용하기 위함. 저장해놓지 않으면 SessionDisconnectEvent에서 사용할 수 없음)
+            headerAccessor.getSessionAttributes().put("chatRoomId", chatRoomId);
+            sendChatMessage(chatRoomId, email, ChatMessageRequest.MessageType.JOIN);
 
         }
 
@@ -69,20 +59,52 @@ public class WebSocketEventListener {
         StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
 
         //헤더에서 채팅방을 나간 유저의 email과 chatRoomId를 추출
-        AuthUser authUser = (AuthUser) Objects.requireNonNull(headerAccessor.getSessionAttributes()).get("authUser");
-        String email = authUser.getEmail();
+        String email = getEmailFromSession(headerAccessor);
         Long chatRoomId = (Long) headerAccessor.getSessionAttributes().get("chatRoomId");
 
-        //메세지 전송
-        if(email != null) {
-
-            ChatMessageRequest chatMessage = new ChatMessageRequest();
-            chatMessage.changeType(ChatMessageRequest.MessageType.LEAVE); //LEAVE타입 설정
-            chatMessage.changeSender(email);
-
-            messagingTemplate.convertAndSend("/topic/"+chatRoomId, chatMessage);
-
+        if (email != null && chatRoomId != null) {
+            sendChatMessage(chatRoomId, email, ChatMessageRequest.MessageType.LEAVE);
         }
+
+    }
+
+    /**
+     * 세션에서 이메일을 가져오는 private 메서드
+     */
+    private String getEmailFromSession(StompHeaderAccessor headerAccessor) {
+
+        return Optional.ofNullable(headerAccessor.getSessionAttributes())
+                .map(attrs -> (AuthUser) attrs.get("authUser"))
+                .map(AuthUser::getEmail)
+                .orElse(null);
+
+    }
+
+    /**
+     * 헤더에서 chatRoomId를 가져오는 private 메서드
+     */
+    private Long getChatRoomIdFromHeader(StompHeaderAccessor headerAccessor) {
+
+        try {
+            return Optional.ofNullable(headerAccessor.getFirstNativeHeader("chatRoomId"))
+                    .map(Long::parseLong)
+                    .orElse(null);
+        } catch (NumberFormatException e) {
+            return null;
+        }
+
+    }
+
+    /**
+     * 채팅 메시지를 생성하고 전송하는 private 메서드
+     */
+    private void sendChatMessage(Long chatRoomId, String email, ChatMessageRequest.MessageType messageType) {
+
+        ChatMessageRequest chatMessage = new ChatMessageRequest();
+        chatMessage.changeType(messageType);
+        chatMessage.changeSender(email);
+
+        messagingTemplate.convertAndSend("/topic/" + chatRoomId, chatMessage);
 
     }
